contents_list.py: use Paragraph class
Rather than wrapping text directly in list items, convert string items
into Paragraphs and make use of the Paragraph type to output wrapped
markdown.

As a result this also fixes the indentation on subsequent lines in the
markdown to properly line up. (they were falling short by one space).
The indentation is also adjusted to align the indentation to match the
guidelines set in Google's Markdown Style Guide [1]. This is two spaces
for numbered lists and three spaces for bullet lists. And in general,
four spaces for nested lists.

This also prepares lists for embedding content beyond pure text.

[1]: https://google.github.io/styleguide/docguide/style.html