document.py: add support for links
Provide a function to create a link in the Document (add_link) which
returns a DocumentLink that can be embedded in the text strings of the
markdown document. Casting the DocumentLink to string generates the
appropriate markdown text to include/use.

At markdown generation time, these link targets will be embedded as
footnote anchors in the document file.