contents_lists.py: add support for nested lists
Allow nesting lists within lists. This commit reworks some of the
markdown generation to support that (item numbering is determined now
based on type of item rather than position in the list because nested
lists should not affect the item numbering).

The examples are updated to match and demonstrate this new feature.